+++ src/Core/ZakYip.WheelDiverterSorter.Core/Sorting/Events/RoutePlannedEventArgs.cs
@@ -13,7 +13,7 @@ public readonly record struct RoutePlannedEventArgs
     /// <summary>
     /// 目标格口ID
     /// </summary>
-    public required int TargetChuteId { get; init; }
+    public required long TargetChuteId { get; init; }
 
     /// <summary>
     /// 路径规划时间（UTC）
